Change of the location of the helper function
