+++ tests/ttnn/unit_tests/operations/pool/test_rotate.py
@@ -7,7 +7,34 @@ import torch
 import ttnn
 from loguru import logger
 from models.common.utility_functions import skip_for_blackhole
-from tests.ttnn.utils_for_testing import assert_with_pcc, get_rotate_tolerances
+from tests.ttnn.utils_for_testing import assert_with_pcc
+
+
+def get_rotate_tolerances(input_shape, angle, interpolation_mode):
+    """
+    Get appropriate tolerances for rotate operation based on tensor size, angle, and interpolation mode.
+
+    The tolerances are calibrated to match the numerical accuracy of the rotate operation,
+    which uses fixed-point arithmetic (Q16.16) for coordinate calculations. Bilinear mode
+    uses the same tolerance as grid_sample operation.
+
+    Args:
+        input_shape (tuple): The shape of the input tensor (N, H, W, C).
+        angle (float): The rotation angle in degrees.
+        interpolation_mode (str): Either "nearest" or "bilinear".
+
+    Returns:
+        tuple: A tuple containing (atol, rtol) tolerances for torch.allclose comparison.
+    """
+    is_diagonal_rotation = angle in [45, 135, -45, -135]
+
+    if interpolation_mode == "nearest":
+        if is_diagonal_rotation:
+            return 6.0, 0.01
+        else:
+            return 5.0, 0.01
+    else:
+        return 1.0, 0.01
 
 
 # ============================================================================
